client-py/requirements.txt
grpcio==1.51.3
grpcio-tools==1.51.3
protobuf==4.21.12
